scripts/requirements.txt
# ─────────────────────────────────────────────
# TSMC 2330 30 天趨勢預測系統 — 依賴套件
# Python 3.10+  |  資料來源：PostgreSQL 17
# ─────────────────────────────────────────────

# ── PostgreSQL 連線（必裝）──────────────────────────────────
psycopg2-binary>=2.9.9

# ── 資料處理（必裝）────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── 技術指標（擇一）────────────────────────────────────────
# ta：純 Python，支援所有版本（推薦，pandas-ta 已無法安裝）
ta>=0.11.0
# pandas-ta：若日後恢復可用則優先使用
# pandas-ta>=0.3.14b0

# ── 機器學習（必裝）────────────────────────────────────────
scikit-learn>=1.4.0
joblib>=1.3.0

# ── 梯度提升樹（必裝）─────────────────────────────────────
xgboost>=2.0.0
lightgbm>=4.3.0
# catboost：Python 3.14 尚不支援，Python 3.10-3.12 可安裝
catboost>=1.2.0

# ── TFT（選裝，無 GPU 亦可用 CPU）─────────────────────────
# 安裝前請先確認 Lightning 版本相容性：
#   pip install "pytorch-forecasting>=1.1" "lightning>=2.1,<2.5"
# 若無 GPU：
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
lightning>=2.1.0
pytorch-forecasting>=1.1.0

# ── 可解釋性（[P3 啟用] 第三輪審查建議）────────────────────
# 用於 feature importance / model 解釋性分析
shap>=0.44.0

# ── 模型追蹤（[P3 啟用] 第三輪審查建議）────────────────────
# 取代手寫 metrics_registry.json，正確的 model versioning 解決方案
mlflow>=2.10.0

# ── 資料漂移監控（[P3 啟用] 第三輪審查建議）────────────────
# 取代手工 PSI 計算，提供更穩健的 reference distribution 比較
evidently>=0.4.0

# ── 監控儀表板（dashboard.py 必裝）────────────────────────
streamlit>=1.30.0
plotly>=5.18.0

# ── 工具 ────────────────────────────────────────────────────
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
